added camera package
