+++ app/src/main/java/vn/mbm/phimp/me/opencamera/Camera/CamObject.java
@@ -1,4 +1,4 @@
-package vn.mbm.phimp.me.opencamera.CamUtil;
+package vn.mbm.phimp.me.opencamera.Camera;
 
 /**
  * Created by manuja on 30/5/17.
