+++ app/src/main/java/vn/mbm/phimp/me/opencamera/Camera/HDRProcessorException.java
@@ -1,4 +1,4 @@
-package vn.mbm.phimp.me.opencamera.CamUtil;
+package vn.mbm.phimp.me.opencamera.Camera;
 
 /** Exception for HDRProcessor class.
  */
