+++ app/src/main/java/vn/mbm/phimp/me/opencamera/Camera/MyWidgetProvider.java
@@ -1,4 +1,4 @@
-package vn.mbm.phimp.me.opencamera.CamUtil;
+package vn.mbm.phimp.me.opencamera.Camera;
 
 import android.app.PendingIntent;
 import android.appwidget.AppWidgetManager;
