+++ app/src/main/java/vn/mbm/phimp/me/opencamera/Camera/PreferenceKeys.java
@@ -1,4 +1,4 @@
-package vn.mbm.phimp.me.opencamera.CamUtil;
+package vn.mbm.phimp.me.opencamera.Camera;
 
 /** Stores all of the string keys used for SharedPreferences.
  */
