+++ app/src/main/java/vn/mbm/phimp/me/opencamera/UI/DrawPreview.java
@@ -19,12 +19,12 @@ import android.view.Surface;
 import android.view.View;
 import android.widget.ImageButton;
 
-import vn.mbm.phimp.me.opencamera.CamUtil.CameraActivity;
+import vn.mbm.phimp.me.opencamera.Camera.CameraActivity;
 import vn.mbm.phimp.me.opencamera.CameraController.CameraController;
-import vn.mbm.phimp.me.opencamera.CamUtil.GyroSensor;
-import vn.mbm.phimp.me.opencamera.CamUtil.MyApplicationInterface;
-import vn.mbm.phimp.me.opencamera.CamUtil.MyDebug;
-import vn.mbm.phimp.me.opencamera.CamUtil.PreferenceKeys;
+import vn.mbm.phimp.me.opencamera.Camera.GyroSensor;
+import vn.mbm.phimp.me.opencamera.Camera.MyApplicationInterface;
+import vn.mbm.phimp.me.opencamera.Camera.MyDebug;
+import vn.mbm.phimp.me.opencamera.Camera.PreferenceKeys;
 import vn.mbm.phimp.me.opencamera.Preview.Preview;
 import vn.mbm.phimp.me.R;
