+++ app/src/main/res/layout/activity_camera.xml
@@ -10,7 +10,7 @@
     android:layout_height="match_parent"
     android:background="@color/black"
     android:layout_weight="5"
-    tools:context=".opencamera.CamUtil.CameraActivity">
+    tools:context=".opencamera.Camera.CameraActivity">
 
     <FrameLayout
         android:id="@+id/preview"
